Extract DiagnosticSuppressor from MetaSyntheticPackageFragmentProvider
